Lot of changes, new equipment mode
Lot of changes, new equipment mode

+++ TESReloaded/KeyboardManager.h
@@ -6,8 +6,8 @@ class KeyboardManager // Never disposed
 public:
 	KeyboardManager();
 
-	bool	OnKeyDown(UInt8 KeyCode);
-	bool	OnKeyUp(UInt8 KeyCode);
+	bool	OnKeyDown(UInt16 KeyCode);
+	bool	OnKeyUp(UInt16 KeyCode);
 	bool	OnMouseDown(UInt8 ButtonIndex);
 	bool	OnMouseUp(UInt8 ButtonIndex);
